fix config server and add admin server and eureka

+++ centre-service/src/main/java/org/projet/centreservice/dtos/TerrainAssignmentEvent.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class TerrainAssignmentEvent {
-    private Long terrainId;
+    private int terrainId;
     private int centreId;
     private String assignmentStatus; // Example: ASSIGNED, FAILED
 }
